refactor: Firebase Remote data source
Data 모듈에서 Firebase 에 대한 의존도를 삭제하고,
app 모듈에서 주입 받아 유연하게 동작하는 코드로 리팩터링
