+++ app/mobile/src/main/java/kr/mastre/jaytv/impl/FirebasePlayItemDto.kt
@@ -1,0 +1,12 @@
+package kr.mastre.jaytv.impl
+
+import kotlinx.serialization.Serializable
+import kr.mastre.playlist.data.PlayableDto
+
+@Serializable
+internal data class FirebasePlayItemDto(val uri: String, val thumb: String) : PlayableDto {
+
+    override val rawUri: String get() = uri
+    override val thumbnail: String get() = thumb
+
+}
