+++ app/mobile/src/main/java/kr/mastre/jaytv/impl/GetPlayListResponse.kt
@@ -1,0 +1,6 @@
+package kr.mastre.jaytv.impl
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+internal data class GetPlayListResponse(val playlist: List<FirebasePlayItemDto>)
