+++ playlist/data/src/main/java/kr/mastre/playlist/data/PlayableDto.kt
@@ -1,0 +1,5 @@
+package kr.mastre.playlist.data
+
+import kr.mastre.playlist.Playable
+
+interface PlayableDto : Playable
