+++ playlist/data/src/main/java/kr/mastre/playlist/data/PlaylistRemoteDataSource.kt
@@ -1,0 +1,8 @@
+package kr.mastre.playlist.data
+
+import io.reactivex.rxjava3.core.Single
+
+interface PlaylistRemoteDataSource {
+
+    fun fetchPlayList() : Single<List<PlayableDto>>
+}
